move dumps functions to scheduler
