+++ distributed/tests/test_executor.py
@@ -26,8 +26,7 @@ from distributed.core import rpc, dumps, loads
 from distributed.client import WrappedKey
 from distributed.executor import (Executor, Future, CompatibleExecutor, _wait,
         wait, _as_completed, as_completed, tokenize, _global_executor,
-        default_executor, _first_completed, ensure_default_get, futures_of,
-        _maybe_complex, dumps_function, dumps_task)
+        default_executor, _first_completed, ensure_default_get, futures_of)
 from distributed.scheduler import Scheduler
 from distributed.sizeof import sizeof
 from distributed.utils import ignoring, sync, tmp_text
@@ -2401,34 +2400,3 @@ def test_dont_delete_recomputed_results(s, w):
         yield gen.sleep(0.01)
 
     yield e._shutdown()
-
-
-def test_maybe_complex():
-    assert not _maybe_complex(1)
-    assert not _maybe_complex('x')
-    assert _maybe_complex((inc, 1))
-    assert _maybe_complex([(inc, 1)])
-    assert _maybe_complex([(inc, 1)])
-    assert _maybe_complex({'x': (inc, 1)})
-
-
-def test_dumps_function():
-    a = dumps_function(inc)
-    assert loads(a)(10) == 11
-
-    b = dumps_function(inc)
-    assert a is b
-
-    c = dumps_function(dec)
-    assert a != c
-
-
-def test_dumps_task():
-    d = dumps_task((inc, 1))
-    assert set(d) == {'function', 'args'}
-
-    f = lambda x, y=2: x + y
-    d = dumps_task((apply, f, (1,), {'y': 10}))
-    assert loads(d['function'])(1, 2) == 3
-    assert loads(d['args']) == (1,)
-    assert loads(d['kwargs']) == {'y': 10}
